Revert "Refactor header parsers (#217)"
This reverts commit 3d1e7d9a7c2a7bee128d33c49c1a76e4ace55e67.
