+++ lib/api/index.js
@@ -9,8 +9,6 @@ const path = require("path"),
   LibhoneyImpl = require("./libhoney"),
   MockImpl = require("./mock");
 
-const { honeycomb, aws } = propagation;
-
 let apiImpl;
 
 function mapFieldsToApp(map) {
@@ -51,23 +49,11 @@ module.exports = {
     return tracker.getTracked();
   },
 
+  TRACE_HTTP_HEADER: propagation.TRACE_HTTP_HEADER,
+  AMAZON_TRACE_HTTP_HEADER: propagation.AMAZON_TRACE_HTTP_HEADER,
   REQUEST_ID_HTTP_HEADER: propagation.REQUEST_ID_HTTP_HEADER,
-  marshalTraceContext: honeycomb.marshalTraceContext,
-
-  // deprecated. Use provider-specific functions and constants
-  TRACE_HTTP_HEADER: honeycomb.TRACE_HTTP_HEADER,
-  AMAZON_TRACE_HTTP_HEADER: aws.TRACE_HTTP_HEADER,
+  marshalTraceContext: propagation.marshalTraceContext,
   unmarshalTraceContext: propagation.unmarshalTraceContext,
-
-  honeycomb: {
-    TRACE_HTTP_HEADER: honeycomb.TRACE_HTTP_HEADER,
-    unmarshalTraceContext: honeycomb.unmarshalTraceContext,
-  },
-
-  aws: {
-    TRACE_HTTP_HEADER: aws.TRACE_HTTP_HEADER,
-    unmarshalTraceContext: aws.unmarshalTraceContext,
-  },
 
   startTrace(metadataContext, withTraceId, withParentSpanId, withDataset) {
     return apiImpl.startTrace(metadataContext, withTraceId, withParentSpanId, withDataset);
