+++ lib/instrumentation/trace-util.js
@@ -21,16 +21,6 @@ const getValueFromHeaders = (requestHeaders, headers) => {
   return undefined;
 };
 
-// adds trace source field to the context and returns the updated context
-function _includeSource(context, key) {
-  if (!context) {
-    return {};
-  }
-  return Object.assign({}, context, {
-    source: `${key} http header`,
-  });
-}
-
 exports.getTraceContext = (traceIdSource, req) => {
   if (typeof traceIdSource === "undefined" || typeof traceIdSource === "string") {
     let headers =
@@ -43,29 +33,52 @@ exports.getTraceContext = (traceIdSource, req) => {
       return {};
     }
     let { value, header } = valueAndHeader;
-    let parsed = {};
 
     switch (header) {
-      //honeycomb trace header
-      case api.honeycomb.TRACE_HTTP_HEADER: {
-        header = api.honeycomb.TRACE_HTTP_HEADER;
-        parsed = api.honeycomb.unmarshalTraceContext(value);
-        return _includeSource(parsed, header);
+      case api.TRACE_HTTP_HEADER: {
+        let parsed = api.unmarshalTraceContext(value);
+        if (!parsed) {
+          return {};
+        }
+        return Object.assign({}, parsed, {
+          source: `${header} http header`,
+        });
       }
 
-      case api.aws.TRACE_HTTP_HEADER: {
-        header = api.aws.TRACE_HTTP_HEADER;
-        parsed = api.aws.unmarshalTraceContext(value);
-        return _includeSource(parsed, header);
+      case api.AMAZON_TRACE_HTTP_HEADER: {
+        let traceId, parentSpanId;
+
+        const split = value.split(";");
+        for (const s of split) {
+          const [name, value] = s.split("=");
+          if (name === "Root") {
+            traceId = value;
+          } else if (name === "Parent") {
+            parentSpanId = value;
+          }
+        }
+
+        if (!traceId) {
+          // if we didn't even get a 'Root=' clause, bail.
+          return {};
+        }
+
+        if (!parentSpanId) {
+          parentSpanId = traceId;
+        }
+
+        return {
+          traceId,
+          parentSpanId,
+          source: `${header} http header`,
+        };
       }
 
       default: {
-        return _includeSource(
-          {
-            traceId: value,
-          },
-          header
-        );
+        return {
+          traceId: value,
+          source: `${header} http header`,
+        };
       }
     }
   } else {
