+++ lib/propagation.js
@@ -1,11 +1,12 @@
 /* global require, exports, Buffer, __dirname */
 const path = require("path"),
-  pkg = require(path.join(__dirname, "..", "..", "package.json")),
+  pkg = require(path.join(__dirname, "..", "package.json")),
   debug = require("debug")(`${pkg.name}:propagation`),
-  schema = require("../schema");
+  schema = require("./schema");
 
-const TRACE_HTTP_HEADER = "X-Honeycomb-Trace";
-exports.TRACE_HTTP_HEADER = TRACE_HTTP_HEADER;
+exports.TRACE_HTTP_HEADER = "X-Honeycomb-Trace";
+exports.AMAZON_TRACE_HTTP_HEADER = "X-Amzn-Trace-Id";
+exports.REQUEST_ID_HTTP_HEADER = "X-Request-ID";
 
 const VERSION = "1";
 exports.VERSION = VERSION;
@@ -26,11 +27,8 @@ exports.VERSION = VERSION;
 //
 // ex: X-Honeycomb-Trace: 1;trace_id=weofijwoeifj,parent_id=owefjoweifj,context=SGVsbG8gV29ybGQ=
 
-// marshalTraceContext takes a trace context object
-// and returns a serialized honeycomb trace header
-exports.marshalTraceContext = marshalTraceContext;
-
-function marshalTraceContext(context) {
+exports.marshalTraceContext = marshalTraceContextv1;
+function marshalTraceContextv1(context) {
   let traceId = context.id;
   let spanId = context.stack[context.stack.length - 1].payload[schema.TRACE_SPAN_ID];
   let contextToSend = context.traceContext || {};
@@ -46,12 +44,8 @@ function marshalTraceContext(context) {
   ).toString("base64")}`;
 }
 
-// unmarshalTraceContext takes a string trace header and returns a context
 exports.unmarshalTraceContext = unmarshalTraceContext;
-
 function unmarshalTraceContext(contextStr) {
-  // this trim and split just splits out the version from the remaining trace header
-  // payload is trace header minus version
   let [version, payload] = contextStr.trim().split(";");
   switch (version) {
     case "1":
@@ -107,10 +101,5 @@ function unmarshalTraceContextv1(payload) {
     }
   }
 
-  return {
-    traceId,
-    parentSpanId,
-    customContext,
-    dataset,
-  };
+  return { traceId, parentSpanId, customContext, dataset };
 }
