+++ lib/propagation.test.js
@@ -1,10 +1,9 @@
 /* global require describe test expect */
-const propagation = require("."),
-  schema = require("../schema"),
-  Span = require("../api/span"),
+const propagation = require("./propagation"),
+  schema = require("./schema"),
+  Span = require("./api/span"),
   cases = require("jest-in-case");
 
-const { honeycomb } = propagation;
 // this context structure is the same as what the libhoney event api implementation generate.
 let testContext = {
   id: "abcdef123456",
@@ -20,7 +19,7 @@ let testContext = {
 describe("marshaling", () => {
   test("version string prefix", () => {
     expect(
-      propagation.marshalTraceContext(testContext).startsWith(`${honeycomb.VERSION};`)
+      propagation.marshalTraceContext(testContext).startsWith(`${propagation.VERSION};`)
     ).toBeTruthy();
   });
 });
